+++ app/src/main/java/com/example/todoapp/data/local/SharedPreferencesManager.kt
@@ -1,4 +1,4 @@
-package com.example.todoapp.data
+package com.example.todoapp.data.local
 
 import android.content.SharedPreferences
 import java.util.UUID
