+++ app/src/main/java/com/example/todoapp/di/component/ActivityComponent.kt
@@ -1,6 +1,6 @@
 package com.example.todoapp.di.component
 
-import com.example.todoapp.ActivityViewModel
+import com.example.todoapp.ui.activity.viewmodel.ActivityViewModel
 import com.example.todoapp.di.scope.ActivityScope
 import dagger.Subcomponent
