+++ app/src/main/java/com/example/todoapp/di/component/AddEditTaskFragmentComponent.kt
@@ -1,7 +1,7 @@
 package com.example.todoapp.di.component
 
 import com.example.todoapp.di.scope.AddEditScope
-import com.example.todoapp.ui.AddEditTaskViewModel
+import com.example.todoapp.ui.fragments.addEdit.viewmodel.AddEditTaskViewModel
 import dagger.Subcomponent
 
 @Subcomponent
