+++ app/src/main/java/com/example/todoapp/di/component/TasksFragmentComponent.kt
@@ -1,7 +1,7 @@
 package com.example.todoapp.di.component
 
 import com.example.todoapp.di.scope.TasksScope
-import com.example.todoapp.ui.TasksViewModel
+import com.example.todoapp.ui.fragments.tasks.viewmodel.TasksViewModel
 import dagger.Subcomponent
