+++ app/src/main/java/com/example/todoapp/ui/activity/viewmodel/ErrorUiState.kt
@@ -1,0 +1,6 @@
+package com.example.todoapp.ui.activity.viewmodel
+
+data class ErrorUiState(
+    val isErrorShown: Boolean = false,
+    val error: Throwable? = null
+)
