+++ app/src/main/java/com/example/todoapp/ui/fragments/addEdit/viewmodel/AddEditUiState.kt
@@ -1,0 +1,10 @@
+package com.example.todoapp.ui.fragments.addEdit.viewmodel
+
+import com.example.todoapp.data.model.Importance
+import java.util.Date
+
+data class AddEditUiState(
+    val text: String = "",
+    val importance: Importance = Importance.COMMON,
+    val deadline: Date? = null,
+)
