+++ app/src/main/java/com/example/todoapp/ui/fragments/tasks/adapter/TodoItemDiffCalculator.kt
@@ -1,7 +1,7 @@
-package com.example.todoapp.ui.adapters
+package com.example.todoapp.ui.fragments.tasks.adapter
 
 import androidx.recyclerview.widget.DiffUtil
-import com.example.todoapp.ui.TodoItemUiState
+import com.example.todoapp.ui.fragments.tasks.viewmodel.TodoItemUiState
 
 class TodoItemDiffCalculator : DiffUtil.ItemCallback<TodoItemUiState>() {
     override fun areItemsTheSame(
