+++ app/src/main/java/com/example/todoapp/ui/fragments/tasks/viewmodel/TasksUiState.kt
@@ -1,0 +1,8 @@
+package com.example.todoapp.ui.fragments.tasks.viewmodel
+
+data class TasksUiState(
+    val todoItemsList: List<TodoItemUiState> = emptyList(),
+    val completedTodoItemsNumber: Int = 0,
+    val showUncompletedItems: Boolean = false,
+    val showError: Boolean = false
+)
