+++ app/src/main/res/layout/fragment_tasks.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="?attr/back_primary"
-    tools:context=".ui.fragments.TasksFragment">
+    tools:context=".ui.fragments.tasks.TasksFragment">
 
     <com.google.android.material.appbar.AppBarLayout
         android:id="@+id/appBarLayout"
