+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/tasksFragment"
-        android:name="com.example.todoapp.ui.fragments.TasksFragment"
+        android:name="com.example.todoapp.ui.fragments.tasks.TasksFragment"
         android:label="fragment_all_tasks"
         tools:layout="@layout/fragment_tasks">
         <action
@@ -22,7 +22,7 @@
     </fragment>
     <fragment
         android:id="@+id/addEditTaskFragment"
-        android:name="com.example.todoapp.ui.fragments.AddEditTaskFragment"
+        android:name="com.example.todoapp.ui.fragments.addEdit.AddEditTaskFragment"
         android:label="fragment_task"
         tools:layout="@layout/fragment_add_edit_task">
         <argument
