+++ app/src/main/res/values/strings.xml
@@ -1,7 +1,5 @@
 <resources>
     <string name="app_name">ToDo App</string>
-    <!-- TODO: Remove or change this placeholder text -->
-    <string name="hello_blank_fragment">Hello blank fragment</string>
     <string name="add_new">Новое</string>
     <string name="done_amount">Выполнено - %1$d</string>
     <string name="my_tasks">Мои задачи</string>
